Add skip_OK to CHECK_VALUE, v1.02
If skip_OK is a true value, the CHECK_VALUE call will not register an OK
message when the value is fine.  This was initially done for stuff like
check_file_stats, which does lots of size, ownership and age checks
against porentially lots of files.  If everything is OK, we only care
that there were X files.