ITM-2651 - Add TEST_CHROOT support to RUN (via shell)
Now, plugin tests will be able to redirect RUN to run binaries in a
different location, even if they are absolute.  It's not a real chroot,
mind you, just a way of shoehorning a path prefix to all executables.

To make it work, the TEST_PLUGINS environment variable must be set to a
non-empty, non-zero value (usually 1) and the TEST_CHROOT var must be
the path to an existent directory.  If both of these preconditions are
met, each RUN call (via shell) will have $ENV{TEST_CHROOT} prepended to
the command to run.