ITM-3541 - Update MPF to make Net::SNMP optional
This fixes a deployment problem on Ecelerity servers, which stupidly
ship with their own RPMs that claim to provide stuff like
'perl(Net::SNMP)', but don't put them in the standard system locations.

Since we only need this dependency on monitoring core boxen (or anywhere
that has tools-hammer-throw-server installed) that's ok.